requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
flask==2.3.3
sqlalchemy==2.0.23
gunicorn==21.2.0
python-dotenv==1.0.0
2captcha-python==1.2.1
psutil==5.9.6
schedule==1.2.1
anthropic==0.9.0
pyyaml==6.0.1
openai==1.3.0
